requirements.txt
fastapi~=0.116.1
uvicorn[standard]
pydantic~=2.11.7
langchain~=0.3.27
langchain-ollama
langchain-openai~=0.3.28
openai~=1.97.1
pydantic-settings~=2.10.1
typing_extensions~=4.14.1
langgraph~=0.6.2
ipython~=9.4.0
langchain-core~=0.3.72
qdrant-client~=1.15.1
typing_extensions~=4.14.1
requests~=2.32.4
typing_extensions~=4.14.1
langchain-text-splitters~=0.3.9
typing_extensions~=4.14.1
langchain-qdrant
starlette~=0.47.2
typing_extensions~=4.14.1
passlib~=1.7.4
python-jose~=3.5.0
python-multipart
pypdf
cryptography>=3.1

aiohttp>=3.8,<4.0
multidict>=4.5,<7.0
langchain-community==0.3.27